Refactor brokers into addon modules
reafactor the broker interface so that each broker is dynamically loaded
as an addon/plugin to impulse.
- Create "api" subproject to hold common API objects for brokers and
  other addons
- Create "app" subproject to hold Impulse implementation code
- Create "docker-broker" subproject to implement the docker broker as an
  addon to Impulse
- Cleanup code style warnings
- Configure Dokka/Kdoc for project documentation generation
- Add Doc strings to public interfaces
- Add Guice based injection for broker factory and config providers
- Auto load broker plugins from config dir on startup
- Dynamically parse broker specific config
- Add dynamic registration event
- Update main package to be combined distro
- Add subproject packages
