+++ app/src/main/templates/club/arson/impulse/BuildConstants.java
