+++ buildSrc/settings.gradle.kts
@@ -1,0 +1,18 @@
+rootProject.name = "buildSrc"
+
+pluginManagement {
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+    }
+}
+
+@Suppress("UnstableApiUsage")
+dependencyResolutionManagement {
+    repositoriesMode.set(RepositoriesMode.PREFER_SETTINGS)
+
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+    }
+}
