+++ buildSrc/src/main/kotlin/dokka-convention.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("org.jetbrains.dokka")
+    id("org.jetbrains.dokka-javadoc")
+}
+
+dokka {
+    dokkaSourceSets.configureEach {
+    }
+}
