--- src/main/kotlin/club/arson/impulse/server/ServerStatus.kt
@@ -1,8 +1,0 @@
-package club.arson.impulse.server
-
-enum class ServerStatus {
-    RUNNING,
-    STOPPED,
-    REMOVED,
-    UNKNOWN
-}
